+++ apps/backoffice-v2/src/Router/Router.tsx
@@ -23,6 +23,10 @@ import { TransactionMonitoring } from '@/pages/TransactionMonitoring/Transaction
 import { TransactionMonitoringAlertsAnalysisPage } from '@/pages/TransactionMonitoringAlertsAnalysis/TransactionMonitoringAlertsAnalysis.page';
 import { Profiles } from '@/pages/Profiles/Profiles.page';
 import { Individuals } from '@/pages/Profiles/Individuals/Individuals.page';
+import { MerchantMonitoring } from '@/pages/MerchantMonitoring/MerchantMonitoring.page';
+import { MerchantMonitoringCreateCheckPage } from '@/pages/MerchantMonitoringCreateCheck/MerchantMonitoringCreateCheck.page';
+import { MerchantMonitoringBusinessReport } from '@/pages/MerchantMonitoringBusinessReport/MerchantMonitoringBusinessReport.page';
+import { MerchantMonitoringLayout } from '@/domains/business-reports/components/MerchantMonitoringLayout/MerchantMonitoringLayout';
 
 const router = createBrowserRouter([
   {
@@ -69,6 +73,27 @@ const router = createBrowserRouter([
             element: <Locale />,
             errorElement: <RouteError />,
             children: [
+              {
+                element: <MerchantMonitoringLayout />,
+                errorElement: <RouteError />,
+                children: [
+                  {
+                    path: '/:locale/merchant-monitoring',
+                    element: <MerchantMonitoring />,
+                    errorElement: <RouteError />,
+                  },
+                  {
+                    path: '/:locale/merchant-monitoring/:businessReportId',
+                    element: <MerchantMonitoringBusinessReport />,
+                    errorElement: <RouteError />,
+                  },
+                  {
+                    path: '/:locale/merchant-monitoring/create-check',
+                    element: <MerchantMonitoringCreateCheckPage />,
+                    errorElement: <RouteError />,
+                  },
+                ],
+              },
               {
                 path: '/:locale/case-management',
                 element: <CaseManagement />,