+++ apps/backoffice-v2/src/lib/blocks/variants/WebsiteMonitoringBlocks/WebsiteMonitoringBlocks.tsx
@@ -1,14 +1,15 @@
 import { ProcessTracker } from '@/common/components/molecules/ProcessTracker/ProcessTracker';
 import { TWorkflowById } from '@/domains/workflows/fetchers';
 import { cells } from '@/lib/blocks/create-blocks-typed/create-blocks-typed';
-import { useWebsiteMonitoringReportBlock } from '@/lib/blocks/variants/WebsiteMonitoringBlocks/hooks/useWebsiteMonitoringReportBlock/useWebsiteMonitoringReportBlock';
+import { useWebsiteMonitoringReportPDFBlock } from '@/lib/blocks/variants/WebsiteMonitoringBlocks/hooks/useWebsiteMonitoringReportPDFBlock/useWebsiteMonitoringReportPDFBlock';
 import { useCasePlugins } from '@/pages/Entity/hooks/useCasePlugins/useCasePlugins';
 import { useCurrentCaseQuery } from '@/pages/Entity/hooks/useCurrentCaseQuery/useCurrentCaseQuery';
 import { BlocksComponent } from '@ballerine/blocks';
+import { NoBlocks } from '@/lib/blocks/components/NoBlocks/NoBlocks';
 
 export const WebsiteMonitoringBlocks = () => {
-  const blocks = useWebsiteMonitoringReportBlock();
-  const { data: workflow } = useCurrentCaseQuery();
+  const blocks = useWebsiteMonitoringReportPDFBlock();
+  const { data: workflow, isLoading } = useCurrentCaseQuery();
   const plugins = useCasePlugins({ workflow: workflow as TWorkflowById });
   const processes = ['merchant-monitoring'];
 
@@ -20,6 +21,7 @@ export const WebsiteMonitoringBlocks = () => {
       <BlocksComponent blocks={blocks} cells={cells}>
         {(Cell, cell) => <Cell {...cell} />}
       </BlocksComponent>
+      {!isLoading && !blocks?.length && <NoBlocks />}
     </div>
   );
 };