+++ sdks/workflow-node-sdk/src/lib/workflow-node-sdk-instance.ts
@@ -8,8 +8,11 @@ export class WorkflowNodeSDKInstance {
     this.#__service = createWorkflow(options);
   }
 
-  subscribe(callback: Parameters<TCreateWorkflowCoreReturn['subscribe']>[0]) {
-    this.#__service.subscribe(callback);
+  subscribe(
+    eventName: Parameters<TCreateWorkflowCoreReturn['subscribe']>[0],
+    callback: Parameters<TCreateWorkflowCoreReturn['subscribe']>[1],
+  ) {
+    this.#__service.subscribe(eventName, callback);
   }
 
   async sendEvent(event: Parameters<TCreateWorkflowCoreReturn['sendEvent']>[0]) {