+++ services/workflows-service/src/business-report/business-report.service.ts
@@ -86,4 +86,16 @@ export class BusinessReportService {
       [projectId],
     );
   }
+
+  async findById<T extends Omit<Prisma.BusinessReportFindFirstOrThrowArgs, 'where'>>(
+    id: string,
+    projectIds: TProjectIds,
+    args?: Prisma.SelectSubset<T, Omit<Prisma.BusinessReportFindFirstOrThrowArgs, 'where'>>,
+  ) {
+    return await this.businessReportRepository.findById(id, projectIds, args);
+  }
+
+  async updateById(...args: Parameters<BusinessReportRepository['updateById']>) {
+    return await this.businessReportRepository.updateById(...args);
+  }
 }