+++ services/workflows-service/src/business/business.service.ts
@@ -46,8 +46,16 @@ export class BusinessService {
     return await this.repository.findById(id, args, projectIds);
   }
 
-  async getByCorrelationId(correlationId: string, projectids: TProjectIds) {
-    return await this.repository.findByCorrelationId(correlationId, {}, projectids);
+  async getByIdUnscoped(id: string, args: Parameters<BusinessRepository['findByIdUnscoped']>[1]) {
+    return await this.repository.findByIdUnscoped(id, args);
+  }
+
+  async getByCorrelationId(
+    correlationId: string,
+    projectids: TProjectIds,
+    args?: Parameters<BusinessRepository['findByCorrelationId']>[2],
+  ) {
+    return await this.repository.findByCorrelationId(correlationId, projectids, args);
   }
 
   async updateById(id: string, args: Parameters<BusinessRepository['updateById']>[1]) {