+++ apps/backoffice-v2/CHANGELOG.md
@@ -1,5 +1,46 @@
 # @ballerine/backoffice-v2
 
+## 0.7.13
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.9.8
+  - @ballerine/workflow-browser-sdk@0.6.14
+  - @ballerine/workflow-node-sdk@0.6.14
+
+## 0.7.12
+
+### Patch Changes
+
+- @ballerine/workflow-browser-sdk@0.6.13
+- @ballerine/workflow-node-sdk@0.6.13
+
+## 0.7.11
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.9.7
+  - @ballerine/ui@0.5.3
+  - @ballerine/workflow-browser-sdk@0.6.12
+  - @ballerine/workflow-node-sdk@0.6.12
+
+## 0.7.10
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/workflow-browser-sdk@0.6.11
+  - @ballerine/workflow-node-sdk@0.6.11
+
+## 0.7.9
+
+### Patch Changes
+
+- @ballerine/workflow-browser-sdk@0.6.10
+- @ballerine/workflow-node-sdk@0.6.10
+
 ## 0.7.8
 
 ### Patch Changes